move: TextBox 합성 컴포넌트 폴더 분리
