+++ src/common/components/Input/Timer/utils/formatTimer.ts
